#1325 - edit form submits
